Move types/ within src/ so it's compiled
Keeping "types" separated from "source code" does not make much sense when enums in TypeScript are both.
